refactor: split DetailsPage into different components
